+++ CHANGELOG.md
@@ -30,6 +30,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Fix "check length methods" from the verifier returning wrong length in case of an incomplete check list.
 - Fix portable repositories generating an invalid prebuild repository structure.
 - Fix writable checks by ignoring the readonly attribute on Windows.
+- Fix package not found message when metadata cannot be parsed.
 
 
 ## [v0.0.3](https://github.com/pack-it/packit/compare/0.0.2...0.0.3) - 2026-07-11